rename 'wrap' to 'dispatch' to better indicate function
