+++ examples/dev/plot_2d_edge_meshes.py
@@ -1,7 +1,7 @@
 import matplotlib.pyplot as plt
 from matplotlib.patches import Polygon
 
-from napari.layers.shapes._accelerated_triangulate_wrap import (
+from napari.layers.shapes._accelerated_triangulate_dispatch import (
     generate_2D_edge_meshes,
 )
