+++ napari/layers/shapes/_accelerated_triangulate_dispatch.py
